added middleware to do the htmx redirect on a redirect to action. Also moved teebox crud to teebox controller
