+++ FairwayFinder/src/FairwayFinder.Web/Areas/Course/Views/Course/ViewCourse.cshtml
@@ -38,7 +38,7 @@
                         <h3>Tees</h3>
                     </div>
                     <div class="col-auto">
-                        <a class="btn btn-primary" asp-action="AddTee" asp-route-courseId="@Model.Course.course_id">
+                        <a class="btn btn-primary" asp-controller="Teebox" asp-action="AddTee" asp-route-courseId="@Model.Course.course_id">
                             <i class="fa-solid fa-plus pe-2"></i> Add Tee Box
                         </a>
                     </div>
@@ -63,7 +63,7 @@
                         {
                             <tr>
                                 <td>
-                                    <a asp-action="ViewTeebox"
+                                    <a asp-controller="Teebox" asp-action="ViewTeebox"
                                        asp-route-courseId="@tee.course_id"
                                        asp-route-teeboxId="@tee.teebox_id">
                                         @tee.teebox_name
